requirements.txt
blinker==1.5
captcha==0.4
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
Flask==2.2.2
Flask-Mail==0.9.1
fonttools==4.43.1
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.8.0
mysql-connector-python==8.0.32
numpy==1.26.1
packaging==23.2
Pillow==9.4.0
protobuf==3.20.3
PyAudio==0.2.13
pyparsing==3.1.1
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
qt5-applications==5.15.2.2.3
razorpay==1.4.1
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
urllib3==2.0.6
Werkzeug==2.2.2
zipp==3.11.0
